Added supplies to the Accounting network
So that we have something to set with the rules, I created supplies out of accounts on Ruedi:
- Completed transfers from Allocatable Flow to each Ruedi contract account. There were about 20 missing transfers.
- Added supplies from contract storage accounts to a single downstream Contract account
- Added supplies from most other accounts to a single downstream Flow account

Made all contract accounts and supplies a cyan color.

Added an account group that is not aggregated at a high level. Turn on the "Disaggregated Temp Group" to see all of the accounts on Ruedi.  Turn it off to show the aggregated view.