Implemented a crude allocation to Phase I and II contracts on Ruedi.
Implemented an allocation algorithm using two rules (25 and 26) to move Allocatable Flow to contracts based on Phase I or Phase II.
It allocates to Phase I first and then once those accounts are full (accrual > limit) , then it allocates proportionally to Phase II.

I broke up the "Contract" water type into "Contract_I" for the four phase 1 accounts and the rest use "Contract" although this can be broken down further once we get the details of the phase defined better.

This seems to work for the test period.