Added a slot for Fish Pool priority order
I added a slot: "AccountingData.Fish Pool Release Priority" that contains
a row for each of the 6 Fish Pools. The values in the table are the priority or order of the releases.
They currently match the 2020 spreadsheet column order in the Input tab.

This assumes the priorities won't change within a run. If they do change, we will need to move this to a series slot.
No rules use this logic yet.