Added supplies to Fish pools from appropriate Contract pools
This commit adds supplies from contract accounts to Fish pools. The
water is transferred to the contract accounts based on Round I or II and then
will be transferred to Fish pools for certain accounts (CWCB, Grand... ) and
when water is leased (Ute, Garfield, Parachute) or donated (Exxon).
This commit sets up the supplies and represents possible transfers that
occurred in 2019-2022.

No rules have been written to do the transfers, yet.