Added Accounts for Fill Rights and Free River
This commit adds two new accounts to Ruedi
- Fill Rights - this is used to track when the reservoir can fill. That
is the First Fill, Second Fill, and Free River Fill.
- Free River - this storage account is used to hold any water
that can be stored if there is excess water and all other water
rights are satisfied.

I also created the supplies from Fill Rights to Free River and from
Fill Rights to the 52 contract accounts.

Finally, I change the FWS Fish Pool to "River District Fish Pool" as that
is more specific.  All of the Fish pools are managed by FWS.