New accounts for First Fill, Second Fill, Free River, and OTR.
This commit implements accounting structure described in the design
document with storage accounts on Ruedi for:
First Fill
Second Fill
OTR (Owe the River)
Distributor

There is also now a Discretionary and Contract account downstream.
All the new accounts are connected by the appropriate supplies.

I had to disable a rules 27-38 to get it to load.
The next step will be getting those to work.