Changes for allocation on Jan 1 and Start of WY
This commit does the following to match the design document in terms of Jan 1 and start of water year:
Changed run range to Init = 12/31/2005; Start 1/1/2006; End 4/30/2007
Changed acct range to Begin 12/31/2005; End 4/30/2007
Changed begin accrual date to 1/1
Added Init rule to set Storage Account Accrual and Storage to zero on init timestep. The res will be re-allocated on Jan1.
Added a policy group “Jan 1 and start of WY Allocation”. This has the following rules to do the distribution/allocation on Jan 1 and the start of the water year, around April 1:
- Added a rule to set Begin Year Allocation for Pools (WT=None) equal to previous timestep storage on Jan 1.  This mimic carryover all, but keeps the right accrual.
- Added a rule to set Distributor Begin Year Allocation to the sum of Contract storage on Dec 31. This is then re-allocated.
- Added rules to set Slot Inflow to negative of previous Accrual to zero out accrual on start of WY.
- Added a rule to set GainLoss to negative of the storage to zero out any storage on start of WY.
- Duplicated the 6 Allocation rules to allocate only the Jan 1 and start of WY water.
- Added a rule to set Slot Inflow to previous Accrual to maintain accrual on Jan 1 for Fill accts.  Also added a rule to set GainLoss to negative of the accrual to maintain Acct storage on Jan 1.

Changed Ruedi Distributor. Transfers Out and Storage convergence to 0.0001% as it was not converging correctly with small changes.
Changed Fish Pools to have no carryover, like the contract accounts.

I believe it is now allocating existing storage correctly on Jan 1 and April 1.  I still need to set up the variable WY start to replace April 1.
I have not addressed reconciliation at all.