Updates to use Start of Fill date
This commit now uses the Start of Fill Date instead of April 1.
The Start of Fill is set by rule 54 on "Ruedi Reservoir.Start of Fill Date".

I did a search and replace of all the instances of April 1 and replaced them with the new function IsStartOfFillDate.
Before the date is set (on April 1), the function returns False.