Intermediate update - added supplies, modified Max Accrual, added rules for initial accrual
This commit has 3 main changes:
1. I added supplies from each Ruedi contract to the downstream Discretionary passthrough.
2. Added rules to set Maximum Accrual for each of the Pool accounts and Contracts.
I then modified the rules to look at the max accrual instead of the accrual limit.
This will allow us to adjust the Max Accrual later when they can refill.
3. Added some logic to compute the initial accrual for the First Fill account.
This looks at how high the historical Pool Elevation got in the previous year.
If it didn't fill, then the First Fill likely didn't happen and the accrual is set accordingly.
A RBS rule also sets the Slot Inflow and GainLoss to get the correct Accrual on Jan 1
as it is the start of the run, but not the start of the WY.