Initial implementation of fish exchanges/leases
The lease/exchanges will be specified in a table slot,
Accounting Data.LeasesAndExchanges.
This has one row per exchange/lease and three columns of input values:
1. On/Off for debugging purposes,
2. the borrow date
3. the borrow/lease volume.
We could add in flow rates, a time range for the borrow or other constraints if needed.

On the exchange date, a rule sets the borrow volume on a transfer supply
from the contract account, up to the storage in the contract account.
The rest is set on a supply from the Replacement to the Fish pool account.
This becomes the borrow supply in the exchange.
Note, there are multiple parallel supplies from Replacement to the
various Fish pools. These have a specific naming convention, for example:
Ruedi Reservoir Replacement Pool to Ruedi Reservoir CWCB Fish Pool FOR Ute Water Conservation District Exchange

Then, there is a RiverWare exchange that tracks the balance of water borrowed from the Replacement pool.

Another rule then pays back the exchange when the contract gets water
allocated to it in priority.   It sets a supply from the contract to the
Replacement pool limited by the contract storage and the exchange balance.