Continued building out Lease logic and structure
This commit continues to build out the structure and logic for the Lease.
In particular:
- Broke up the Leases and Exchanges table into
-- Lease slot that specifies the name as a row label, On/Off and whether
a borrow from the Replacement pool is allowed.  If not allowed, the
contract must have enough storage at the time of the lease or a warnign is posted.
--  A slot for each lease. It now contains the lease date and lease volume.

- Added in all of the leases from 2015 to 2023 as per the spreadsheet from Tim.
There are 14 different leases with some occurring on multiple dates, for a total of 28 lease events.
The spreadsheet only listed the water year of each lease so I made up day of year that they occur.
At some point, this should be updated to match the historical lease date.

- Added in all the supplies and exchanges as appropriate.
- Updated the rule logic for this new structure

I Tested by running from 2006-2021 and made sure all leases were triggered as appropriate

Next step is to add logic to reduce Contract Initial Request when a lease occurs.