Reduction in contract demand when Lease occurs
This commit improves the lease logic as follows:
- when a lease occurs, the Initial Request for the rest of the CY is
reduced proportionally by the lease amount.  Many accounts that lease
don't actually have enough demand, so in most cases, the
Initial request either becomes zero (or stays zero).
- when a lease occurs before the April X WY start, it must carryover into that WY.
The rule that sets Max Accrual now looks at the total lease amount
from Jan 1 to the start of WY and reduces the contract's accrual limit by that amount

This was tested on a few leases, but more testing with realistic lease dates is needed.