Added a Power Right account and updated the logic
To better track the power right, I added a passthrough account named
Power Right. It gets water from the allocatable flow account and then
releases it downstream back to the allocatable flow chain.

I also updated the logic in rule 42 that sets the first, second, free river storable flows.
Now it also sets the "FlowToPassForPowerRight"  and then uses that in the computation for 2nd and free river fill.

The reservoir is reconciled from a high level view but there are some differences if you zoom way into the plot .

Finally, I updated the Ruedi Reconcilation plot with this new account and better axis scaling .