Decreased Convergence on Fish accounts and fixed SumFishStorage function
First, there was a reconciliation difference of about 1 or 2 af on some timesteps. Inflows and outflows all match, so I decreased convergence on Fish.Storage accounts and it fixed most of the differences. This is a common issue in accounting models; when a rule changes an account outflow by a small amount, the relative storage change is within convergence and is not reset.  I would change all storage convergence values to be 100 times smaller: 0.0001%. I only did the Fish account as it is a pain as they have be changed one at a time. In 9.2 it will be easier to change the convergence of acct slots in the global slot config, so we should change them all once we move to 9.2.

After making that change, I found another issue on 10/15/2010, where there is an input Fish release, but it thinks it is in Winter so won’t release it.  I tracked it down and found that the SumFishStorage function was not actually doing a sum, but grabbing the last value.  This error made the model think it was in Winter, when it was really still Fall.  See IsWinter function at the bottom for the post-10/15 logic.  With the change to SumFishStorage, it now stays Fall through October when there is Fish Storage remaining.

Finally, I added a Ruedi.StorageReconciliation expr slot to compute the difference between acct storage and physical storage. This should be very close to zero to be reconciled. With both changes, it is nearly zero (within convergence) for the entire run.